# Add buffer transfer optimization to gpu_to_comp pass 