--- a/pmlc/conversion/gpu_to_comp/BUILD
+++ b/pmlc/conversion/gpu_to_comp/BUILD
@@ -20,6 +20,23 @@ gentbl(
     ],
 )
 
+gentbl(
+    name = "enums-gen",
+    tbl_outs = [
+        (
+            "-gen-enum-decls",
+            "enums.h.inc",
+        ),
+        (
+            "-gen-enum-defs",
+            "enums.cc.inc",
+        ),
+    ],
+    tblgen = "@llvm-project//mlir:mlir-tblgen",
+    td_file = "enums.td",
+    td_srcs = ["@llvm-project//mlir:OpBaseTdFiles"],
+)
+
 plaidml_cc_library(
     name = "gpu_to_comp",
     srcs = [
@@ -31,6 +48,7 @@ plaidml_cc_library(
     ],
     deps = [
         ":gen-pass",
+        ":enums-gen",
         "//pmlc/dialect/comp/ir",
         "@llvm-project//mlir:GPUDialect",
         "@llvm-project//mlir:IR",