@jbruestle 2020-12-02
Looks good overall.  Two high level points:  1) It would be nice to have a separate PR for the GPU->COMP lowering optimizations and the Vulkan changes, however I tried to pull out just the Vulkan changes and that broke, not sure why exactly (see https://github.com/plaidml/plaidml/pull/1585).  I'm OK landing these as one PR if we need to, but I'd like to understand why better.  2) The changes in GPU->COMP lowering are basically mixing optimization and conversion/lowering into one block of code.  In general the design is more composable if we make conversions as simple as possible and then make optimizations happen afterwards (or before).  Again, I don't think that should necessarily block this PR, but just wanted to comment for future reference.