@YangleiZouIntel 2020-12-02
> Looks good overall. Two high level points: 1) It would be nice to have a separate PR for the GPU->COMP lowering optimizations and the Vulkan changes, however I tried to pull out just the Vulkan changes and that broke, not sure why exactly (see #1585). I'm OK landing these as one PR if we need to, but I'd like to understand why better. 2) The changes in GPU->COMP lowering are basically mixing optimization and conversion/lowering into one block of code. In general the design is more composable if we make conversions as simple as possible and then make optimizations happen afterwards (or before). Again, I don't think that should necessarily block this PR, but just wanted to comment for future reference.

Thanks Jeremy for the note. 

For (2) I have split the code in this PR in two passes (one before gpu_to_comp and one after gpu_to_comp) and some changes in gpu_to_comp pass, please see https://github.com/plaidml/plaidml/pull/1590 . I will add a separate PR for vulkan. 

For (1) the current Vulkan and OpenCL backends have different comp representation and I changed gpu_to_comp pass in this PR so the vulkan code in #1585 itself will broke if the changes in gpu_to_comp pass in this PR is not applied. With the changes in gpu_to_comp pass, the current Vulkan and OpenCL backends will share the same comp dialect representation, only the buffer memory space and execEnv will be different.