@YangleiZouIntel 2020-12-08
Closed since this PR is splited into https://github.com/plaidml/plaidml/pull/1590  and https://github.com/plaidml/plaidml/pull/1605